add Gate::authorize

+++ app/Http/Controllers/HotelController.php
@@ -6,6 +6,7 @@ use App\Http\Resources\HotelResource;
 use App\Http\Resources\UsersInRoomsInHotel;
 use App\Models\Hotel;
 use App\Models\Room;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 
@@ -83,6 +84,7 @@ class HotelController extends Controller
      */
     public function destroy(Hotel $hotel)
     {
+        Gate::authorize('delete', $hotel);
         $hotel->delete();
 
         return response()->json([
@@ -93,6 +95,9 @@ class HotelController extends Controller
     }
     public function addRoomInHotel(Hotel $hotel, Room $room)
     {
+        Gate::authorize('view', $hotel);
+        Gate::authorize('update', $room);
+
         $room->hotel_id = $hotel->id;
         $room->save();
 
@@ -103,8 +108,12 @@ class HotelController extends Controller
             ]
         ]);
     }
-    public function getRoomsInHotels(Hotel $hotel, Room $room)
+    public function getRoomsInHotels()
     {
+        Gate::authorize('viewAny', Hotel::class);
+        Gate::authorize('viewAny', Room::class);
+        Gate::authorize('viewAny', User::class);
+        
         $hotels = Hotel::with('rooms.users')->get();
 
         return UsersInRoomsInHotel::collection($hotels);
